WIP: notifikasi model
