+++ app/Models/Notifikasi.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Notifikasi extends Model
+{
+    protected $table = 'notifikasi';
+
+    protected $primaryKey = 'id';
+}
