+++ app/Models/Otp.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\Master;
+namespace App\Models;
 
 use App\Traits\UsesUuid;
 use Illuminate\Database\Eloquent\Model;
